requirements.txt
  
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
pyngrok==4.1.5
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
six==1.14.0
twilio==6.38.1
urllib3==1.25.8
Werkzeug==1.0.1
PyMySQL==0.10.1
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.4.4
Flask-SocketIO==4.3.1
eventlet==0.29.1
SQLAlchemy==1.3.18
flask-login==0.5.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
PyAudio==0.2.11
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-speech==2.0.0
googleapis-common-protos==1.52.0